requirements.txt
asgiref==3.4.1
awsebcli==3.20.2
botocore==1.21.65
cement==2.8.2
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.3
Django==3.2.9
django-admin-honeypot==1.1.0
django-ckeditor==6.2.0
django-js-asset==1.2.2
django-session-timeout==0.1.0
future==0.16.0
idna==3.3
jmespath==0.10.0
pathspec==0.5.9
Pillow==8.4.0
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pywin32==303
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
urllib3==1.26.7
wcwidth==0.1.9
